telegram_bot/requirements.txt
aiogram==3.4.1
aiohttp==3.9.1
httpx==0.26.0
openpyxl==3.1.2
python-dotenv==1.0.0
